[CoP]: Реорганизация ядра часть 2(без проблем)
Многие файлы теперь находятся в рендере(как в CoP), а не используются из xrEngine, также удалил некоторые файлы в xrGame - они не нужны. Особенно r2_test_hw, который тупо используется как скриптовый метод, что за дебилизм? Также в device.h удалил методы, с пометкой `//DELETE THIS!!!`, оставленные в a2d741f

P.S: В этой версии команда vid_mode и опции запускаются нормально

+++ trunk/xrGame/ui/UIWindow.h
@@ -35,7 +35,7 @@ public:
 	virtual void			DetachChild			(CUIWindow* pChild);
 	virtual bool			IsChild				(CUIWindow* pChild) const;
 	virtual void			DetachAll			();
-	int						GetChildNum			()								{return m_ChildWndList.size();} 
+	size_t					GetChildNum			()								{return m_ChildWndList.size();} 
 
 	void					SetParent			(CUIWindow* pNewParent);
 	CUIWindow*				GetParent			()	const							{return m_pParentWnd;}
